.travis.yml += TDD testing "make test" feature - i.e. ...
... using "make test" target for automatic unittesting.
In (TDD) Test Driven Development manner, first we make target that fails,
by using function `false`. Later let's check if it will display as failed
on Travis DI dashboard. If yes, then it means that it works! So next step will
be to fix it to `true`, providing skeleton for future use.